+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Follow users by username for status updates.
+- Decode session variables from the auth token.
 
 ### Changed
 - Increase the default socket timeout to 30 seconds.